% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tourTableUi.R
\name{tourBtnToggleDisplay}
\alias{tourBtnToggleDisplay}
\title{tourBtnToggleDisplay}
\usage{
tourBtnToggleDisplay(text)
}
\arguments{
\item{text}{text of the button}
}
\description{
toggle help panel
}

