% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tourTableUi.R
\name{tourBtnToggleEditor}
\alias{tourBtnToggleEditor}
\title{tourBtnToggleEditor}
\usage{
tourBtnToggleEditor(text)
}
\arguments{
\item{text}{text of the button}
}
\description{
toggle help panel
}

