% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tourTableServer.R
\name{tourCheckItems}
\alias{tourCheckItems}
\title{tourCheckItems}
\usage{
tourCheckItems(session, visible = FALSE)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer}

\item{visible}{Select only visible elements.}
}
\value{
input$tourItems values
}
\description{
tourCheckItems check html for tourGroup and input label
}

