% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tourTableServer.R
\name{tourCheckMembers}
\alias{tourCheckMembers}
\title{tourCheckMembers}
\usage{
tourCheckMembers(session, visible = FALSE)
}
\arguments{
\item{session}{The session object passed to function in shinyServer}

\item{visible}{Select only visible elements.}
}
\value{
input$tour_members_list values
}
\description{
tourCheckMembers check html for tourGroup and input label
}

