% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tourTableServer.R
\name{tourConfig}
\alias{tourConfig}
\title{tourConfig}
\usage{
tourConfig(dbPath = "tourItemTable.sqlite", tableName = "tourItems",
  idEditor = "tour_ace_editor", idSelectId = "tour_select_id",
  idSelectGroup = "tour_select_group",
  idCheckHighLight = "tour_check_highlight", disableUpdate = F)
}
\arguments{
\item{dbPath}{path to sqlite items self contained database}

\item{tableName}{name of table containing items info}

\item{idEditor}{id of ShinyAce editor used internaly}

\item{idSelectId}{id of select input for item's id}

\item{idSelectGroup}{id of select input for item's group}
}
\value{
list of configuration to be bassed to tourManager()
}
\description{
Create a list to group configuration arguments for tourManager()
}

