% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tourTableServer.R
\name{tourDbDocTob64}
\alias{tourDbDocTob64}
\title{tourDbDocTo64}
\usage{
tourDbDocTob64(dbPath, dbTableName, disableKnit, selector = NULL)
}
\arguments{
\item{dbPath}{path to sqlite databse}

\item{dbTableName}{table containing tour information}

\item{disableKnit}{disable knit (see tourPreviewRender64)}

\item{selector}{selector id for which to extract doc column content}
}
\description{
extract doc field (stored in ascii base64) from db for given selector, knit R code , produce html, convert back to b64
}

