% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tourTableUi.R
\name{tourEditor}
\alias{tourEditor}
\title{tourEditor}
\usage{
tourEditor(title = "Tour step content edit", idEditor = "tour_ace_editor",
  idSelectGroup = "tour_select_group", idSelectId = "tour_select_id",
  idCheckHighLight = "tour_check_highlight", width = 12)
}
\arguments{
\item{idEditor}{outputId of ace editor}

\item{idSelectGroup}{id of group selectize input}

\item{idSelectID}{id of id selectize input}
}
\description{
Editor for selected step
}

