% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tourTableServer.R
\name{tourFilterId}
\alias{tourFilterId}
\title{tourFilterId}
\usage{
tourFilterId(session, selectedGroup, dbPath = "tour.sqlite",
  tableName = "tourItems", idSelectId = "tourSelectId")
}
\arguments{
\item{session}{current shiny session}

\item{selectedGroup}{group selected}

\item{dbPath}{path to sqlite db}

\item{tableName}{table containing items data}

\item{idSelectId}{id of id select input as defined in tourEditor()}
}
\description{
Given a group of items, update id selectInput
}

