% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tourTableUi.R
\name{tourGroup}
\alias{tourGroup}
\title{tourGroup}
\usage{
tourGroup(idGroup, title, ...)
}
\arguments{
\item{idGroup}{group id. Ex. subsection-intro}

\item{title}{title of the group}

\item{...}{children}
}
\description{
define new group for shiny tour
}

