% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tourTableServer.R
\name{tourHighlightItem}
\alias{tourHighlightItem}
\title{tourHighlightItem}
\usage{
tourHighlightItem(session, itemSelector, enable = T)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer}

\item{itemSelector}{Item to highlight}
}
\description{
Toggle red box around item selected
}

