% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tourTableServer.R
\name{tourHighlightMember}
\alias{tourHighlightMember}
\title{tourHighlightMember}
\usage{
tourHighlightMember(session, memberSelector, enabled)
}
\arguments{
\item{session}{The session object passed to function in shinyServer}

\item{memberSelector}{Member to highlight}
}
\description{
Toggle red box around member selected
}

