% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tourTableServer.R
\name{tourInit}
\alias{tourInit}
\title{tourInit}
\usage{
tourInit(session = NULL, members = NULL, dbPath = "tour.sqlite",
  dbTableName = "tourMembers", idSelectGroup = "tourSelectGroup",
  disableUpdate = F)
}
\arguments{
\item{session}{shiny session}

\item{members}{list of members from input$tour_members_list (generated by shinyTour.js)}

\item{dbPath}{path to SQLite db}

\item{dbTableName}{name of SQLite db table that contains members info}

\item{idSelectGroup}{id of group select input defined in tourEditor()}

\item{disableUpdate}{avoid update}
}
\description{
If nothing found, it will ask tourCheckMembers to look again at html for new members.
}

