% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tourTableServer.R
\name{tourItemsManager}
\alias{tourItemsManager}
\title{tourManager}
\usage{
tourItemsManager(input, session, config)
}
\arguments{
\item{config}{configuration list of class "tourConfig"}
}
\description{
Manage shinyTour
}

