% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tourTableServer.R
\name{tourItemsToDb}
\alias{tourItemsToDb}
\title{tourItemsToDb}
\usage{
tourItemsToDb(dbCon, itemsTable, tableName = "tourItems")
}
\arguments{
\item{dbCon}{database connection}

\item{itemsTable}{table of items and group generated by tourTable}

\item{tableName}{name of table in sqlite db}
}
\description{
Write news items to existing DB table, create new if needed.
}

