% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tourTableServer.R
\name{tourMembersManager}
\alias{tourMembersManager}
\title{tourManager}
\usage{
tourMembersManager(input = NULL,
  session = shiny::getDefaultReactiveDomain(), config = NULL)
}
\arguments{
\item{config}{configuration list of class "tourConfig"}
}
\description{
Manage shinyTour
}

