% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tourTableServer.R
\name{tourMembersToDb}
\alias{tourMembersToDb}
\title{tourMembersToDb}
\usage{
tourMembersToDb(dbPath, membersTable, dbTableName = "tourMembers")
}
\arguments{
\item{membersTable}{table of members and group generated by tourTable}

\item{dbTableName}{name of table in sqlite db}

\item{dbCon}{database connection}
}
\description{
Write news members to existing DB table, create new if needed.
}

