% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tourTableUi.R
\name{tourPanel}
\alias{tourPanel}
\title{tourBtnToggleDisplay}
\usage{
tourPanel(title = "Tour step content edit", idEditor = "tour_ace_editor",
  idSelectGroup = "tour_select_group",
  idSelectSelector = "tour_select_selector", width = 12)
}
\arguments{
\item{idEditor}{outputId of ace editor}

\item{idSelectGroup}{id of group selectize input}

\item{text}{text of the button
tourPanel

Editor for selected step}

\item{idSelectID}{id of id selectize input}
}
\description{
toggle help panel
}

