% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tourTableServer.R
\name{tourPreviewRender64}
\alias{tourPreviewRender64}
\title{tourPreviewRender64}
\usage{
tourPreviewRender64(txt, disableKnit = T)
}
\arguments{
\item{txt}{text/markdown/Rmarkdown to render in html}

\item{disableKnit}{option to disable knit. An option to use standard markdown could be implemented.}
}
\description{
tourPreviewRender64
}

