% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tourTableServer.R
\name{tourTable}
\alias{tourTable}
\title{tourTable}
\usage{
tourTable(tourItems)
}
\arguments{
\item{tourItems}{input$tourItems}
}
\value{
table of tour groups and input
}
\description{
list of items found (input$touritems) to data.frame.
}

