% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tourTableServer.R
\docType{class}
\name{tourConfig}
\alias{tourConfig}
\title{Tour configuration}
\format{An \code{\link{R6Class}} generator object}
\usage{
tourConfig
}
\arguments{
\item{dbPath}{path to sqlite members self contained database}

\item{dbTableName}{name of table containing members info}

\item{idEditor}{id of ShinyAce editor used internaly}

\item{idSelectSelector}{id of select input for member's id}

\item{idSelectGroup}{id of select input for member's group}
}
\description{
create a R6 object to handle tour configuration and store temporary data.
}
\keyword{data}

