% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tourTableServer.R
\name{tourFilterId}
\alias{tourFilterId}
\title{tourFilterId}
\usage{
tourFilterId(session, selectedGroup, dbPath = "tour.sqlite",
  dbTableName = "tourMembers", idSelectSelector = "tourSelectId")
}
\arguments{
\item{session}{current shiny session}

\item{selectedGroup}{group selected}

\item{dbPath}{path to sqlite db}

\item{dbTableName}{table containing members data}

\item{idSelectSelector}{id of selector's select input. in tourEditor()}
}
\description{
Given a group of members, update id selectInput
}

