% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tourTableServer.R
\name{tourInit}
\alias{tourInit}
\title{tourInit}
\usage{
tourInit(session = shiny::getDefaultReactiveDomain(), items = NULL,
  dbPath = "tour.sqlite", tableName = "tourItems",
  idSelectGroup = "tourSelectGroup", disableUpdate = F)
}
\arguments{
\item{session}{shiny session}

\item{items}{list of items from input$tourItems (generated by shinyTour.js)}

\item{dbPath}{path to SQLite db}

\item{tableName}{name of SQLite db table that contains items info}

\item{idSelectGroup}{id of group select input defined in tourEditor()}

\item{disableUpdate}{avoid update}
}
\description{
tourInit
}

