% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tourTableServer.R
\name{tourTable}
\alias{tourTable}
\title{tourTable}
\usage{
tourTable(tourMembers)
}
\arguments{
\item{tourMembers}{input$tour_members_list}
}
\value{
table of tour groups and input
}
\description{
list of members found (input$tourMembers) to data.frame.
}

