% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tourTableServer.R
\name{tourUpdateDoc}
\alias{tourUpdateDoc}
\title{tourUpdateDoc}
\usage{
tourUpdateDoc(selectedId, text, dbPath = "tour.sqlite",
  tableName = "tourItems")
}
\arguments{
\item{selectedId}{item id selected}

\item{text}{text value from editor}

\item{dbPath}{path to sqliteDb}

\item{tableName}{name of table containing items data}
}
\description{
As soon as the doc in editor is modified, update SQLite table
}

